+++ src/cli/global/upgrade_all.rs
@@ -1,7 +1,7 @@
 use clap::Parser;
 use indexmap::IndexMap;
 
-use rattler_conda_types::{MatchSpec, Platform};
+use rattler_conda_types::{MatchSpec, NamedChannelOrUrl, Platform};
 
 use crate::config::{Config, ConfigCli};
 
@@ -20,7 +20,7 @@ pub struct Args {
     /// By default, if no channel is provided, `conda-forge` is used, the channel
     /// the package was installed from will always be used.
     #[clap(short, long)]
-    channel: Vec<String>,
+    channel: Vec<NamedChannelOrUrl>,
 
     #[clap(flatten)]
     config: ConfigCli,
